size_t for GetKmerIndex position

+++ ntcoding.h
@@ -5,5 +5,5 @@ uint32_t NtChar2Int (char nt);
 uint32_t NtChar2IntCaseInsensitive (char nt);
 uint32_t TransitionNt (uint32_t nt);
 void GenerateShapePos(std::string shape);
-uint32_t GetKmerIndexAtPos(char* sequence, uint32_t pos, uint32_t seed_size);
+uint32_t GetKmerIndexAtPos(char* sequence, size_t pos, uint32_t seed_size);
 int IsTransitionAtPos(int t);
